Integrate Configuration & Weather API with Flutter App (#5)
* Implement configuration & weather api with flutter app

* Use melos commands

* Install flutter in GA

* Set flutter version in GA

* Fixing import invalid case

* Updating package basic info

+++ apps/panel/lib/core/models/general/display_configuration.dart
@@ -1,12 +1,12 @@
 class DisplayConfigurationModel {
   final bool _darkMode;
-  final double _brightness;
+  final int _brightness;
   final int _screenLockDuration;
   final bool _screenSaver;
 
   DisplayConfigurationModel({
     required bool darkMode,
-    required double brightness,
+    required int brightness,
     required int screenLockDuration,
     required bool screenSaver,
   })  : _darkMode = darkMode,
@@ -16,7 +16,7 @@ class DisplayConfigurationModel {
 
   bool get hasDarkMode => _darkMode;
 
-  double get brightness => _brightness;
+  int get brightness => _brightness;
 
   int get screenLockDuration => _screenLockDuration;
 
@@ -24,7 +24,7 @@ class DisplayConfigurationModel {
 
   DisplayConfigurationModel copyWith({
     bool? darkMode,
-    double? brightness,
+    int? brightness,
     int? screenLockDuration,
     bool? screenSaver,
   }) {
